marketplace-notifications/requirements.txt
selenium
webdriver-manager
beautifulsoup4
python-dotenv
pydantic
pyyaml
